attempting to refactor to use hook
