+++ src/components/Quote/Quote.js
@@ -1,13 +1,13 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-const Quote = ({ quote, character, image }) => (
+const Quote = ({ quote, character, image }) => {
   <section>
     <h2>{character}</h2>
     <img src={image} />
     <p>{quote}</p>
-  </section>
-);
+  </section>;
+};
 
 Quote.propTypes = {
   quote: PropTypes.string.isRequired,
