--- src/hooks/quotes.js
@@ -1,9 +1,0 @@
-import { useState, useEffect } from 'react';
-
-export const useQuotes = quotes => {
-const [quotes, setQuotes] = useState([]);
-
-const fetchQuotes = () => {
-    getQuotes(quotes).then(quotes => setQuotes(quotes));
-  };
-}
